+++ content-aside.php
@@ -4,7 +4,7 @@
  *
  * @package WordPress
  * @subpackage WP_Forge
- * @since WP-Forge 5.3.3
+ * @since WP-Forge 5.3.4
  */
 ?>
 
@@ -16,6 +16,7 @@
 			<?php if ( comments_open() ) : ?>
 				<span class="genericon genericon-comment"></span> <?php comments_popup_link( '<span class="leave-reply">' . __( 'Comment', 'wpforge' ) . '</span>', __( '1 Comment', 'wpforge' ), __( '% Comments', 'wpforge' ) ); ?>
 			<?php endif; // comments_open() ?>
+			<?php edit_post_link( __( 'Edit', 'wpforge' ), '<span class="edit-link"><span class="genericon genericon-edit"></span> ', '</span>' ); ?>
 		</div><!-- end .entry-meta-header -->
 		<div class="aside">
 			<h1 class="entry-title"><a href="<?php the_permalink(); ?>" title="<?php echo esc_attr( sprintf( __( 'Permalink to %s', 'wpforge' ), the_title_attribute( 'echo=0' ) ) ); ?>" rel="bookmark"><?php the_title(); ?></a></h1>
@@ -25,10 +26,7 @@
 		</div><!-- .aside -->
 
 		<footer class="entry-meta">
-			<div class="entry-meta-footer">
-				<?php wpforge_entry_meta_footer(); ?><br />
-				<?php edit_post_link( __( 'Edit', 'wpforge' ), '<span class="edit-link"><span class="genericon genericon-edit"></span> ', '</span>' ); ?>
-			</div><!-- end .entry-meta-footer -->
+			<?php wpforge_entry_meta_footer(); ?>
 			<?php get_template_part( 'content', 'author' ); ?>
 		</footer><!-- .entry-meta -->
 	</article><!-- #post -->