+++ inc/customizer/customizer.php
@@ -1,5 +1,4 @@
 <?php
-
 /**
 * WP-Forge Theme Customizer
 * A Theme Customizer for WP-Forge. Adds the individual sections, settings, and controls to the theme customizer
@@ -18,7 +17,7 @@
 * @author Devin Price (@devinsays)
 * @see http://wptheming.com/2012/06/add-options-to-theme-customizer-default-sections/
 *
-* @since WP-Forge 5.3.3
+* @since WP-Forge 5.3.4
 */
 
 function wpforge_customizer( $wp_customize ) { // Begin WP-Forge Theme Customizer
@@ -64,7 +63,7 @@ $wp_customize->get_section('static_front_page')->description = __( 'Set up a fro
 		'transport'  => 'postMessage',
 	 ) );
 		$wp_customize->add_control('blogname', array( 
-			'label'    => __('Site Title', 'reactor'),
+			'label'    => __('Site Title', 'wpforge'),
 			'section'  => 'wpforge_header',
 			'priority' => 3,
 		 ) );
@@ -77,7 +76,7 @@ $wp_customize->get_section('static_front_page')->description = __( 'Set up a fro
 		'transport'  => 'postMessage',
 	 ) );
 		$wp_customize->add_control('blogdescription', array( 
-			'label'    => __('Tagline', 'reactor'),
+			'label'    => __('Tagline', 'wpforge'),
 			'section'  => 'wpforge_header',
 			'priority' => 4,
 		 ) );
@@ -94,130 +93,142 @@ $wp_customize->get_section('static_front_page')->description = __( 'Set up a fro
 		'section'  => 'wpforge_header',
 		'type'     => 'checkbox',
 		'priority' => 4,
-	 ));
-	 
-// Navigation Section
-	// Main Menu Position
+	 ));						
+
+// Top-Bar/Off-Canvas Section
+    $wp_customize->add_section('wpforge_off_canvas', array(
+		'title' => __('Top-Bar &amp; Off-Canvas', 'wpforge'),
+		'description' => __('Configure Top-Bar Navigation and Off-Canvas mobile menu.', 'wpforge'),
+		'priority' => 30,
+	));
+	// Top-Bar Position
 	$wp_customize->add_setting('wpforge_nav_position',array(
 		'default' => 'normal',
 		'type'    => 'theme_mod',
 		'capability' => 'manage_options',
-		'priority' => 110,	
+		'priority' => 1,	
 	));
 	$wp_customize->add_control('wpforge_nav_position',array(
 		'type' => 'select',
 		'label' => __('Main Menu Position', 'wpforge'),
-		'section' => 'nav',
+		'section' => 'wpforge_off_canvas',
 		'choices' => array(
-			'normal' => 'Normal Position',
-			'top'    => 'Top of Browser',
-			'sticky'   => 'Contain-To-Grid Sticky',
+			'normal' => __('Nomral Position', 'wpforge'),
+			'top'    => __('Top of Browser - Scroll', 'wpforge'),
+			'fixed'  => __('Top of Browser - Fixed', 'wpforge'),
+			'sticky' => __('Contain-To-Grid-Sticky', 'wpforge'),
 		),
 	));
-	// Main Menu Display Type
-	$wp_customize->add_setting('wpforge_nav_display',array(
-		'default' => 'scroll',
-		'type'    => 'theme_mod',
-		'capability' => 'manage_options',
-			'priority' => 120,		
-	));
-	$wp_customize->add_control('wpforge_nav_display',array(
-		'type' => 'select',
-		'label' => __('If Top of Browser, scroll or fixed?', 'wpforge'),
-		'section' => 'nav',
-		'choices' => array(
-			'scroll' => 'Scroll',
-			'fixed'  => 'Fixed',
-		),
-	));
-	// Main Menu Main Link Anchor
+	// Top-Bar Main Link Anchor
 	$wp_customize->add_setting('wpforge_nav_title',array(
 		'default' => 'no',
 		'type'    => 'theme_mod',
 		'capability' => 'manage_options',
-		'priority' => 130,		
+		'priority' => 2,		
 	));
 	$wp_customize->add_control('wpforge_nav_title',array(
 		'type' => 'select',
 		'label' => __('Change text for Home Link?', 'wpforge'),
-		'section' => 'nav',
-		'choices' => array(
-			'no' => 'No',
-			'yes'  => 'Yes',
+		'section' => 'wpforge_off_canvas',
+		'choices'	=> array(
+			'no' 	=> __('No', 'wpforge'),
+			'yes'  	=> __('Yes', 'wpforge'),
 		),
 	));
-	// Main Menu Main Link Text
+	// Top-Bar Main Link Text
 	$wp_customize->add_setting('wpforge_nav_text',array(
 		'default' => '',
-		'priority' => 140,
+		'priority' => 3,
 	));
 	$wp_customize->add_control('wpforge_nav_text',array(
 		'label' => __('If Yes, Add New Anchor Text','wpforge'),
-		'section' => 'nav',
+		'section' => 'wpforge_off_canvas',
 		'type' => 'text',
 		'sanitize_callback' => 'wpforge_sanitize_navtxt',
-	));	 						
-
-// Off-Canvas Section
-    $wp_customize->add_section('wpforge_off_canvas', array(
-		'title' => __('Off-Canvas', 'wpforge'),
-		'description' => __('Configure Off-Canvas mobile menu.', 'wpforge'),
-		'priority' => 30,
 	));
 	// Use Off-Canvas for Mobile?
 	$wp_customize->add_setting('wpforge_mobile_display',array(
 		'default' => 'no',
 		'type'    => 'theme_mod',
 		'capability' => 'manage_options',
-		'priority' => 10,	
+		'priority' => 4,	
 	));
 	$wp_customize->add_control('wpforge_mobile_display',array(
 		'type' => 'select',
 		'label' => __('Use Off-Canvas for Mobile?', 'wpforge'),
 		'section' => 'wpforge_off_canvas',
 		'choices' => array(
-			'no'	=> 'No',
-			'yes'	=> 'Yes',
+			'no'	=> __('No', 'wpforge'),
+			'yes'	=> __('Yes', 'wpforge'),
 		),
 	));
-	// Off-Canvas Position	
+	// Off-Canvas Hamburger Icon Position	
 	$wp_customize->add_setting('wpforge_mobile_position',array(
 		'default' => 'left',
 		'type'    => 'theme_mod',
 		'capability' => 'manage_options',
-		'priority' => 20,	
+		'priority' => 5,	
 	));
 	$wp_customize->add_control('wpforge_mobile_position',array(
 		'type' => 'select',
 		'label' => __('Display Off-Canvas Left or Right?', 'wpforge'),
 		'section' => 'wpforge_off_canvas',
 		'choices' => array(
-			'left'	=> 'Left',
-			'right'	=> 'Right',
+			'left'	=> __('Left', 'wpforge'),
+			'right'	=> __('Right', 'wpforge'),
 		),
 	));	
 		
-//The Post Section 
-    $wp_customize->add_section('wpforge_posts',array(
-		'title' => __('Posts', 'wpforge'),
-		'description' => __('Modify how posts appear in your theme.', 'wpforge'),
+//The Layout Section
+    $wp_customize->add_section('wpforge_layout',array(
+		'title' => __('Theme Layout', 'wpforge'),
+		'description' => __('Modify how certain aspects of your theme are displayed.', 'wpforge'),
 		'priority' => 40,
     ));
-    // Show full posts or excerpts
+    // Theme Width
+	$wp_customize->add_setting('wpforge_theme_width',array(
+		'default' => '64rem',
+		'type' => 'theme_mod',
+	));
+	$wp_customize->add_control('wpforge_theme_width',array(
+		'label' => __('Theme Width','wpforge'),
+		'section' => 'wpforge_layout',
+		'type' => 'text',
+		'sanitize_callback' => 'wpforge_sanitize_theme_width',
+		'priority' => 1,
+	));
+	// Sidebar Position
+	$wp_customize->add_setting('wpforge_sidebar_position',array(
+		'default' => 'right',
+		'type' => 'theme_mod',
+		'capability' => 'manage_options',		
+		'sanitize_callback' => 'wpforge_sanitize_sidebar_position',
+		'priority' => 2,		
+	));
+	$wp_customize->add_control('wpforge_sidebar_position',array(
+		'type' => 'select',
+		'label' => __('Sidebar Position', 'wpforge'),
+		'section' => 'wpforge_layout',
+		'choices' => array(
+			'right' => __('Right', 'wpforge'),
+			'left' 	=> __('Left', 'wpforge'),
+		),
+	));
+    // Show full post or excerpts
 	$wp_customize->add_setting('wpforge_post_display',array(
 		'default' => 'full',
 		'type' => 'theme_mod',
 		'capability' => 'manage_options',		
 		'sanitize_callback' => 'wpforge_sanitize_post_display',
-		'priority' => 1,		
+		'priority' => 3,		
 	));
 	$wp_customize->add_control('wpforge_post_display',array(
 		'type' => 'select',
 		'label' => __('Show full post or excerpt?', 'wpforge'),
-		'section' => 'wpforge_posts',
+		'section' => 'wpforge_layout',
 		'choices' => array(
-			'full' => 'Full Post',
-			'excerpt' => 'Excerpt',
+			'full' 		=> __('Full Post', 'wpforge'),
+			'excerpt' 	=> __('Excerpt', 'wpforge'),
 		),
 	));
 	// Show thumbnails on index page
@@ -226,15 +237,15 @@ $wp_customize->get_section('static_front_page')->description = __( 'Set up a fro
 		'type' => 'theme_mod',
 		'capability' => 'manage_options',		
 		'sanitize_callback' => 'wpforge_sanitize_thumb_display',
-		'priority' => 2,		
+		'priority' => 4,		
 	));
 	$wp_customize->add_control('wpforge_thumb_display',array(
 		'type' => 'select',
 		'label' => __('Display post thumbnails?', 'wpforge'),
-		'section' => 'wpforge_posts',
+		'section' => 'wpforge_layout',
 		'choices' => array(
-			'no' => 'No',
-			'yes' => 'Yes',
+			'no' 	=> __('No', 'wpforge'),
+			'yes' 	=> __('Yes', 'wpforge'),
 		),
 	));
 	// Show thumbnails on single post view
@@ -243,30 +254,30 @@ $wp_customize->get_section('static_front_page')->description = __( 'Set up a fro
 		'type' => 'theme_mod',
 		'capability' => 'manage_options',		
 		'sanitize_callback' => 'wpforge_sanitize_single_thumb_display',
-		'priority' => 3,		
+		'priority' => 5,		
 	));
 	$wp_customize->add_control('wpforge_single_thumb_display',array(
-		'type' => 'select',
-		'label' => __('Show single post view thumbnail?', 'wpforge'),
-		'section' => 'wpforge_posts',
-		'choices' => array(
-			'no' => 'No',
-			'yes' => 'Yes',
+		'type' 		=> 'select',
+		'label' 	=> __('Show single post view thumbnail?', 'wpforge'),
+		'section' 	=> 'wpforge_layout',
+		'choices' 	=> array(
+			'no' 	=> __('No', 'wpforge'),
+			'yes' 	=> __('Yes', 'wpforge'),
 		),
 	));
 	// Number of posts to display on sitemap page
 	$wp_customize->add_setting('wpforge_sitemap_count',array(
 		'default' => '15',
-		'priority' => 4,
+		'priority' => 6,
 	));
 	$wp_customize->add_control('wpforge_sitemap_count',array(
-		'label' => __('Number of Posts Displayed On Sitemap','wpforge'),
-		'section' => 'wpforge_posts',
+		'label' => __('Site Map Page: Number of Posts to Display','wpforge'),
+		'section' => 'wpforge_layout',
 		'type' => 'text',
 		'sanitize_callback' => 'wpforge_sanitize_sitemaptxt',
 	));	
 				
-//The Footer Section
+// The Footer Section
     $wp_customize->add_section('wpforge_footer',array(
 		'title' => __('Footer','wpforge'),
 		'description' => __('Modify the footer text of your theme.', 'wpforge'),
@@ -276,12 +287,30 @@ $wp_customize->get_section('static_front_page')->description = __( 'Set up a fro
 		'default' => '',
 		'type' => 'theme_mod',
 		'transport' => 'postMessage',
+		'priority' => 1,
 	));
 	$wp_customize->add_control('wpforge_footer_text',array(
 		'label' => __('Copyright Text','wpforge'),
 		'section' => 'wpforge_footer',
 		'type' => 'text',
 		'sanitize_callback' => 'wpforge_sanitize_footer_text',
+	));
+	$wp_customize->add_setting('wpforge_footer_position',array(
+		'default' => 'right',
+		'type'    => 'theme_mod',
+		'capability' => 'manage_options',
+		'priority' => 2,	
+	));
+	$wp_customize->add_control('wpforge_footer_position',array(
+		'type' => 'select',
+		'sanitize_callback' => 'wpforge_sanitize_footer_position',		
+		'label' => __('Copyright Position', 'wpforge'),
+		'section' => 'wpforge_footer',
+		'choices' => array(
+			'right'  	=> __('Text Right - Nav Left', 'wpforge'),
+			'center' 	=> __('Text &amp; Nav Centered', 'wpforge'),
+			'left'   	=> __('Text Left - Nav Right', 'wpforge'),			
+		),
 	));
 
 // Background Section
@@ -374,9 +403,9 @@ $wp_customize->get_section('static_front_page')->description = __( 'Set up a fro
 		'section'    => 'wpforge_background',
 		'type'       => 'select',
 		'choices'    => array(
-			'left'       => __('Left', 'wpforge'),
-			'center'     => __('Center', 'wpforge'),
-			'right'      => __('Right', 'wpforge'),
+			'left'      => __('Left', 'wpforge'),
+			'center'    => __('Center', 'wpforge'),
+			'right'     => __('Right', 'wpforge'),
 			'none'      => __('None', 'wpforge'),
 		),
 	));
@@ -395,7 +424,7 @@ $wp_customize->get_section('static_front_page')->description = __( 'Set up a fro
 		'section'    => 'wpforge_background',
 		'type'       => 'select',
 		'choices'    => array(
-			'scroll'      => __('Scroll', 'wpforge'),
+			'scroll'    => __('Scroll', 'wpforge'),
 			'fixed'     => __('Fixed', 'wpforge'),
 		),
 	));
@@ -467,16 +496,22 @@ $wp_customize->get_section('static_front_page')->description = __( 'Set up a fro
  * Sanitation Section - This is where we add our sanitation functions for text inputs, check boxes, radio buttons and select lists
  * I like to keep them all in one area for organization.
  * 
- * @since WP-Forge 5.3.3
+ * @since WP-Forge 5.3.4
  */
  
- // Header Sanitation Section
+/*
+ * Header Sanitation Section
+ */
+
+ // Sanitize site title
 function wpforge_sanitize_site_title( $input ) {
     return wp_kses_post( force_balance_tags( $input ) );
 }
+// Sanitize site description
 function wpforge_sanitize_site_description( $input ) {
     return wp_kses_post( force_balance_tags( $input ) );
 }
+// Sanitize the option to show or not show header text
 function wpforge_sanitize_headtxt( $input ) {
     if ( $input == 1 ) {
         return 1;
@@ -485,13 +520,32 @@ function wpforge_sanitize_headtxt( $input ) {
     }
 }
 
-// Navigation Sanitation Section
+/*
+ * Navigation Sanitation Section
+ */
+
+// Sanitize the the text input
 function wpforge_sanitize_navtxt( $input ) {
     return wp_kses_post( force_balance_tags( $input ) );
 }
 
-// Post Sanitation Section
-function wpforge_sanitize_post_display( $input ) {
+// Layout Sanitation Section
+function wpforge_sanitize_theme_width( $input ) { //theme width
+    return wp_kses_post( force_balance_tags( $input ) );
+}
+function wpforge_sanitize_sidebar_position( $input ) { //sidebar position
+    $valid = array(
+        'right' => 'Right',
+		'left' 	=> 'Left',
+    );
+	
+    if ( array_key_exists( $input, $valid ) ) {
+        return $input;
+    } else {
+        return '';
+    }
+}
+function wpforge_sanitize_post_display( $input ) { // post display
     $valid = array(
         'full' => 'Full Post',
 		'excerpt' => 'Excerpt',
@@ -535,7 +589,9 @@ function wpforge_sanitize_sitemaptxt( $input ) {
 function wpforge_sanitize_footer_text( $input ) {
     return wp_kses_post( force_balance_tags( $input ) );
 }
-
+function wpforge_sanitize_font_size( $input ) {
+    return wp_kses_post( force_balance_tags( $input ) );
+}
 // Background Sanitation Section
 function wpforge_sanitize_background_repeat( $input ) {
     $valid = array(
@@ -592,7 +648,7 @@ function wpforge_sanitize_background_attachment( $input ) {
  * Add postMessage support for some sections of our Theme Customizer.
  * @param WP_Customize_Manager $wp_customize Theme Customizer object.
  * 
- * @since WP-Forge 5.3.3
+ * @since WP-Forge 5.3.4
  */
 $wp_customize->get_setting( 'blogname' )->transport = 'postMessage';
 $wp_customize->get_setting( 'blogdescription' )->transport = 'postMessage';
@@ -617,13 +673,13 @@ add_action( 'customize_register', 'wpforge_customizer' ); // End WP-Forge Theme 
  * @author Anthony Wilhelm (@awshout)
  * @see https://github.com/awtheme/reactor
  *
- * @since WP-Forge 5.3.3
+ * @since WP-Forge 5.3.4
  */
 function wpforge_customizer_css() {
 	do_action('wpforge_customizer_css');
 		
 	$output = ''; $body_css = '';	
-		
+	
 	if ( get_theme_mod('wpforge_background_color') ) {
 	    $body_css .= 'background-color: ' . get_theme_mod('wpforge_background_color') . ';';
 	}
@@ -646,32 +702,38 @@ function wpforge_customizer_css() {
 	    $body_css .= 'color: ' . get_theme_mod('wpforge_text_color') . ';';
 	}
 	if ( !empty( $body_css ) ) {
-	    $output .= "\n" . 'body { ' .  $body_css . ' }';
+	    $output .= "\n" . 'body {' .  $body_css . '}';
+	}
+	if ( get_theme_mod('wpforge_theme_width') ) {
+	    $output .= "\n" . '#wrapper {max-width: ' . get_theme_mod('wpforge_theme_width') . ';}';
+	}
+	if ( get_theme_mod('wpforge_sidebar_position') == 'left') {
+	    $output .= "\n" . '#content.columns {float: right!important;} #secondary.columns {float: left!important;}';
 	}
 	if ( 0 == get_theme_mod('hide_headtxt', 1) ) {
-	    $output .= "\n" . '.site-title, .site-description { display: none; }';
+	    $output .= "\n" . '.site-title, .site-description {display: none;}';
 	}	
 	if ( get_theme_mod('wpforge_title_color') ) {
-	    $output .= "\n" . 'h1,h2,h3,h4,h5,h6 { color: ' . get_theme_mod('wpforge_title_color') . '; }';
+	    $output .= "\n" . 'h1,h2,h3,h4,h5,h6 {color: ' . get_theme_mod('wpforge_title_color') . ';}';
 	}		
 	if ( get_theme_mod('wpforge_link_color') ) {
-	    $output .= "\n" . 'a { color: ' . get_theme_mod('wpforge_link_color') . '; }';
+	    $output .= "\n" . 'a {color: ' . get_theme_mod('wpforge_link_color') . ';}';
 	}
 	if ( get_theme_mod('wpforge_hover_color') ) {
-	    $output .= "\n" . 'a:hover { color: ' . get_theme_mod('wpforge_hover_color') . '; }';
-	}			
+	    $output .= "\n" . 'a:hover {color: ' . get_theme_mod('wpforge_hover_color') . ';}';
+	}
 	
-	echo ( $output ) ? '<style>' . apply_filters('wpforge_customizer_css', $output) . "\n" . '</style>' . "\n" : '';
+	echo ( $output ) ? '<style type="text/css">' . apply_filters('wpforge_customizer_css', $output) . "\n" . '</style>' . "\n" : '';
 }
 add_action('wp_head', 'wpforge_customizer_css');
 
 /**
  * Registers our theme customizer preview with WordPress.
  *
- * @since WP-Forge 5.3.3
+ * @since WP-Forge 5.3.4
  */
 function wpforge_customize_preview_js() {
-	wp_enqueue_script('wpforge-customizer', get_template_directory_uri() . '/inc/customizer/js/theme-customizer.js', array('jquery', 'customize-preview' ),'1.0.0', true);
+	wp_enqueue_script('wpforge-customizer', get_template_directory_uri() . '/inc/customizer/js/theme-customizer.js', array('jquery', 'customize-preview' ),'5.3.4', true);
 }
 add_action( 'customize_preview_init', 'wpforge_customize_preview_js' );