+++ inc/customizer/js/theme-customizer.js
@@ -3,7 +3,7 @@
  *
  * Contains handlers to make the WP-Forge Theme Customizer preview reload changes asynchronously.
  * Things like site title, description, and background color changes.
- * @since WP-Forge 5.3.3
+ * @since WP-Forge 5.3.4
  */
 
 ( function( $ ) {
@@ -28,9 +28,14 @@
 				$('.site-title, .site-description').css('display', 'block');
 			};
         });
-    });
+    });    
 	// Footer
 	wp.customize('wpforge_footer_text', function(value) {
+		value.bind(function(to) {
+			$('.site-info').html(to);
+		});
+	});	
+	wp.customize('textarea_settings', function(value) {
 		value.bind(function(to) {
 			$('.site-info').html(to);
 		});