+++ js/foundation/foundation.orbit.js
@@ -1,5 +1,5 @@
 /**
- * @since WP-Forge 5.3.3
+ * @since WP-Forge 5.3.4
  */
 ;(function ($, window, document, undefined) {
   'use strict';
@@ -405,7 +405,7 @@
   Foundation.libs.orbit = {
     name: 'orbit',
 
-    version: '5.3.1',
+    version: '5.3.3',
 
     settings: {
       animation: 'slide',